refactor(server): rename server/api -> server/modules (PR 2/5)
Pure rename: server/api/ did not host the HTTP-API layer (that lives in
app.js and, post PR 3, in server/routes/). It holds business-logic handlers
per domain (coach, push, sync, nutrition, ...), so the directory semantics
matches 'modules' much better than 'api'.

Only file moves + import path updates + doc references. Zero behavior
change. Public HTTP endpoints under /api/... are untouched.

A follow-up commit adds this commit's SHA to .git-blame-ignore-revs so
GitHub blame ignores the rename noise.
