+++ server/modules/barcode.js
